Finished the untested code for linearAssignment(GMatrix, ShouldMaximize, epsilon)
Added copyright and license boiler-plate and file comment.

Changed the transformation to apply to non-finite matrix elements as
well.


Moved the code for standardizing a maximization problem into its own
(private) algorithm.  My thought was to reuse it in the brute-force
code.  But thinking about it, the brute force code is so simple that
it is probably better to write two versions of the code since that
will enhance finding problems in the transformation routines when it
comes time to do testing.