Added code to do a brute-force linear assignment problem solving. Added one very simple test-case, which it passes.
Added stream insertion operator (<<) to GSimpleAssignment

Wrote code to calculate the next k-permutation of n elements in the
same way that std::next_permutation calculates the next n-permutation
of n elements.  Right now it is hidden behind an anonymous namespace,
but it should probably be released to the world.  This code is
relatively well tested - its unit test is called from the
testLinearAssignment code.