Fixed bug in operator<< for GSimpleAssignment - if there were assignment pairs to print and the first element of set A was not assigned, a comma would be printed at the beginning.  I fixed this bug to make it only print commas when an assignment pair had already been printed.
Added test case to reproduce the problem described above (before I fixed it)

Added two new status variables to check whether LAPVJRCT had executed
each of the two branches of the if statement labeled h==min.  (The
statement is not if(h==min) because this is testing equality of
floating point numbers, so an epsilon comparison is used instead, so I
only label it as h==min with a comment)

Added two new test cases that use matrices that were designed to
execute the two branches of the h==min if statement.

Modified the Makefile so it builds the fam tool I used for searching
for the test cases.

Modified the findAugMatrix.cpp source to the fam tool so it can search
for the two new matrices.  Added command-line argument parsing.  Made
it easier to change the size of the hard-coded matrix.  Made the tool
stop after finding 10 matrices.  Made the tool print out status every 10,000 searches rather than every 1000 searches.