#included zlib.h since newer versions of libpng's png.h don't include  it by default.  Everything compiles fine for me still.  But I don't  have a windows box to test things on.
This was in response to the following email:

Subject: 	Waffles & new PNG
From: 	Matthew Chambers (matt.chambers42@gmail.com)
To: 	eric_moyer@yahoo.com;
Cc: 	
Bcc: 	
Date: 	Wednesday, December 21, 2011 12:58 PM


Hi Eric,

We're including a subset of the Waffles GClasses in our project and we
already have a version of PNG but it's newer than the one GClasses
uses. The only incompatibility is that the newer PNG versions do not
include zlib.h, so now GImage.cpp needs to do that in addition to
including png.h.

Thanks,
-Matt